+++ src/app/(main)/page.tsx
@@ -4,19 +4,19 @@ import prisma from "@/lib/prisma";
 import { redirect } from "next/navigation";
 import Image from "next/image";
 // Components
-import BackgroundVideo from "@/components/main/BackgroundVideo";
-import TickerTape from "@/components/main/TickerTape";
-import ContactButton from "@/components/main/ContactButton";
-import VideoModal from "@/components/main/modals/VideoModal";
-import HeaderCopy from "@/components/main/HeaderCopy";
-import ScrollChevrons from "@/components/main/ScrollChevrons";
-import ParallaxImage from "@/components/main/ParallaxImage";
-import SegmentText from "@/components/main/SegmentText";
-import SegmentImage from "@/components/main/SegmentImage";
-import NotFound from "@/components/main/NotFound";
+import BackgroundVideo from "@/components/main/header/BackgroundVideo";
+import TickerTape from "@/components/main/shared/TickerTape";
+import ContactButton from "@/components/main/shared/ContactButton";
+import VideoModal from "@/components/main/shared/VideoModal";
+import HeaderCopy from "@/components/main/header/HeaderCopy";
+import ScrollChevrons from "@/components/main/header/ScrollChevrons";
+import ParallaxImage from "@/components/main/segment/ParallaxImage";
+import SegmentText from "@/components/main/segment/SegmentText";
+import SegmentImage from "@/components/main/segment/SegmentImage";
+import NotFound from "@/components/main/shared/NotFound";
 import DataError from "@/components/main/shared/DataError";
-import ContactModal from "@/components/main/modals/ContactModal";
-import CaseStudyModal from "@/components/main/modals/CaseStudyModal";
+import ContactModal from "@/components/main/shared/ContactModal";
+import CaseStudyModal from "@/components/main/caseStudy/CaseStudyModal";
 // Types
 import { ExtendedSegment } from "@/lib/types";
 import { ExtendedPage } from "@/lib/types/pageTypes";
