+++ src/components/dashboard/layout/DesktopSideBar.tsx
@@ -17,7 +17,7 @@ import { signOut } from "next-auth/react";
 // constants
 import { navLinks } from "@/lib/constants";
 // functions
-import { countUnreadMessages } from "@/lib/functions";
+import countUnreadMessages from "@/lib/utils/messageUtils/countUnreadMessages";
 // components
 import DesktopNavLink from "./DesktopNavLink";
 import ChangeAvatarModal from "./ChangeAvatarModal";
