+++ src/components/main/caseStudy/CaseStudyModal.tsx
