+++ src/components/main/caseStudy/ViewCSVideoModal.tsx
