+++ src/components/main/segment/SegmentImage.tsx
