+++ src/components/main/segment/SegmentText.tsx
