+++ src/components/main/shared/TickerTape.tsx
