+++ src/lib/utils/pageUtils/onChevronScroll.ts
@@ -1,0 +1,5 @@
+export default function handleChevronOnScroll(chevron: HTMLDivElement) {
+	if (chevron) {
+		chevron.classList.replace("opacity-100", "opacity-0");
+	}
+}
